Finalizing Timer-Counter Driver
As I move between chapters in the book and their respective code branches, code is being shuffled around, with me manually carrying relevant changes over as I do not not understand enough git or the intentions of the code-base's creators.

I suspect this related to why I'm encountering usage faults after an outdated sleep routine is called. Regardless, I'd like to commit now and attempt to merge changes as this may be the procedure going forward.

(The repository's branches map directly to book chapters, with code for specific areas of the EIE system. The book, however, assumes each chapter's code is fleshed out in sequence. It's not clear to me why. None of this is explicitly stated. I'm just happy to be able to learn from this resource in isolation )