Switching to compatible workspace version
It was necessary to reimage a machine using an old backup which included an older version of IAR Embedded workbench. The project files, stored separately, are not backwards compatible. Fortunately, the EIE repo includes two workspaces/projects, one of which is of a much older  version and this suitable for my purposes after minor tweaks to workspace configuration.